Mostly final version of the app (for GWC)
It's still buggy near the end of the app lifetime in which level = 12
because 3 icons (take quiz, recommendations, and vocab list) can't
handle an index that overshoots array length. It's also buggy if the
user scores a 0/11 on the diagnostic, repeating the diagnostic quiz and
not allowing the user to pass to level 1 unless he scores a 5 or
greater. However, the middle segments are working just fine and are the
best representation of how the app should work.